backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
databases[postgresql]
alembic
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
pydantic
openfga-sdk
httpx
pytest
pytest-asyncio
bandit
